impl: Create Viewer component.
Moved the viewer from the main App.js to the viewer component to make it routeable.
