+++ frontend/flashfolio/src/Viewer.js
@@ -1,0 +1,25 @@
+import React, {useState, useEffect} from "react";
+import Flashcard from "./Flashcard";
+
+export default function Viewer() {
+
+	const [flashcard, setFlashcard] = useState("");
+
+	useEffect(() => {
+
+		const reqOpt = {
+			method: 'post',
+			headers: {'Content-Type': 'application/json'},
+			body: {'ID': 0}
+		}
+
+		fetch('http://localhost:1337/getDeck', reqOpt)
+			.then(resp => resp.json())
+			.then(data => setFlashcard(data))
+
+	}, [])
+
+	return (
+			<Flashcard flashcard={flashcard} />
+	);
+}
